move application settings cache
